update script for gift-card
